refactor: reduce amount of account summary components, fix every TS and data problem
