+++ package.json
@@ -11,6 +11,7 @@
 	"dependencies": {
 		"@cyntler/react-doc-viewer": "^1.17.0",
 		"@hookform/resolvers": "^4.1.0",
+		"@radix-ui/react-accordion": "^1.2.11",
 		"@radix-ui/react-checkbox": "^1.1.5",
 		"@radix-ui/react-dialog": "^1.1.7",
 		"@radix-ui/react-label": "^2.1.2",
