+++ src/components/account/create-account-component.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import CreateAccountForm from './create-account-form'
-import CreateAccountSummary from './create-account-summary'
+import CreateAccountForm from './forms/create-account-form'
+import AccountSummaryFromStore from './summaries/account-summary-store'
 
 const CreateAccountComponent = () => {
 	return (
 		<div className="w-4/5 mx-auto max-w-[500px] lg:grid grid-cols-2 lg:max-w-[1000px] lg:gap-20 text-primary mb-16 lg:mb-0">
-			<CreateAccountSummary />
+			<AccountSummaryFromStore />
 			<CreateAccountForm />
 		</div>
 	)
