+++ src/components/account/edit-account-component.tsx
@@ -1,16 +1,26 @@
 import React from 'react'
 import { ClientDataType } from '@/lib/types'
-import EditAccountForm from './edit-account-form'
-import AccountSummary from './account-summary'
+import EditAccountForm from './forms/edit-account-form'
+import AccountSummaryFromDB from './summaries/account-summary-db'
 
 const EditAccountComponent = async ({
 	clientData,
 }: {
 	clientData: ClientDataType
 }) => {
+	const clientSelection = {
+		typeOfVehicle: clientData.vehicles_type || [],
+		typeOfVehicleContract: clientData.vehicles_contract || [],
+		chargingStations: clientData.charging_stations_type || [],
+		chargingStationsContract: clientData.charging_stations_contract || [],
+		pif: clientData.pif || [],
+		deployment: clientData.deployment || [],
+		project: clientData.project || [],
+	}
+
 	return (
 		<div className="w-4/5 mx-auto max-w-[500px] lg:grid grid-cols-2 lg:max-w-[1000px] lg:gap-20 text-primary mb-16 lg:mb-0">
-			<AccountSummary clientData={clientData} />
+			<AccountSummaryFromDB clientSelection={clientSelection} />
 			<EditAccountForm clientData={clientData} />
 		</div>
 	)
