+++ src/components/account/forms/edit-account-form.tsx
@@ -1,11 +1,11 @@
 'use client'
 
 import { useToast } from '@/hooks/use-toast'
-import { CreateAccountType } from '@/lib/types'
-import { createAccountSchema } from '@/lib/zod-schemas'
+import { ClientDataType, UpdateAccountType } from '@/lib/types'
+import { updateAccountSchema } from '@/lib/zod-schemas'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { usePathname, useRouter } from 'next/navigation'
-import React, { useState } from 'react'
+import { useRouter } from 'next/navigation'
+//import { useState } from 'react'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import {
 	Form,
@@ -14,97 +14,62 @@ import {
 	FormItem,
 	FormLabel,
 	FormMessage,
-} from '../ui/form'
-import { Input } from '../ui/input'
-import { Button } from '../ui/button'
-import { LuEyeClosed, LuEye } from 'react-icons/lu'
-import { useStore } from '@/store/store'
-import { useShallow } from 'zustand/shallow'
-import { signUpClient } from '@/actions/clients'
-import { selectionArrayFromStoreToDB } from '@/lib/utils'
+} from '../../ui/form'
+import { Input } from '../../ui/input'
+import { Button } from '../../ui/button'
+//import { LuEyeClosed, LuEye } from 'react-icons/lu'
+/* import { useStore } from '@/store/store'
+import { useShallow } from 'zustand/shallow' */
+import { clientUpdate } from '@/actions/clients'
 
-const CreateAccountForm = () => {
-	const [isView, setIsView] = useState(false)
-	const [isViewConfirm, setIsViewConfirm] = useState(false)
+const EditAccountForm = ({ clientData }: { clientData: ClientDataType }) => {
+	//const [isView, setIsView] = useState(false)
+	//const [isViewConfirm, setIsViewConfirm] = useState(false)
 
 	const { toast } = useToast()
 
 	const router = useRouter()
-	const pathname = usePathname()
 
-	console.log('pathname', pathname)
-
-	const { storeSector, storeData } = useStore(
+	/* const { storeSector, storeData, storeGeographies } = useStore(
 		useShallow((state) => ({
 			storeSector: state.sector,
 			storeData: state.data,
+			storeGeographies: state.geographies,
 		}))
-	)
+	) */
 
-	const form = useForm<CreateAccountType>({
-		resolver: zodResolver(createAccountSchema),
+	const form = useForm<UpdateAccountType>({
+		resolver: zodResolver(updateAccountSchema),
 		defaultValues: {
-			name: '',
-			familyName: '',
-			orgName: '',
-			email: '',
-			password: '',
-			confirmPassword: '',
+			name: clientData.name,
+			familyName: clientData.family_name,
+			orgName: clientData.org_name ?? undefined,
+			email: clientData.email,
 		},
 	})
 
 	const isSubmitting = form.formState.isSubmitting
 
-	const submitHandler: SubmitHandler<CreateAccountType> = async (
-		data: CreateAccountType
+	const submitHandler: SubmitHandler<UpdateAccountType> = async (
+		data: UpdateAccountType
 	) => {
-		console.log('formatting data', data)
-		const fullData = {
-			...data,
-			sector: storeSector.value,
-			vehicles_type: selectionArrayFromStoreToDB(
-				storeData.eMobility.typeOfVehicle
-			),
-			vehicles_contract:
-				storeData.eMobility.typeOfVehicleContract?.map(
-					(item) => item.value
-				) || [],
-			charging_stations_type: selectionArrayFromStoreToDB(
-				storeData.eMobility.chargingStations
-			),
-			charging_stations_contract:
-				storeData.eMobility.chargingStationsContract?.map(
-					(item) => item.value
-				) || [],
-			pif: selectionArrayFromStoreToDB(storeData.eMobility.pif),
-			deployment: selectionArrayFromStoreToDB(
-				storeData.eMobility.deployment
-			),
-			project: selectionArrayFromStoreToDB(storeData.eMobility.project),
-		}
+		const updateEmail = data.email !== clientData.email
 
 		try {
-			let response
-			if (pathname.includes('confirm-account')) {
-				// If we are on the confirm-account page, we need to pass the id
-				const id = pathname.split('/').pop()
-				if (!id) {
-					throw new Error('No ID found in the URL')
-				}
-				response = await signUpClient(fullData, id)
-			} else {
-				// Otherwise, we just call the signUpClient without an id
-				response = await signUpClient(fullData)
-			}
+			const response = await clientUpdate(
+				data,
+				clientData.id,
+				updateEmail
+			)
 
 			if (response) {
 				toast({
 					title: 'Thank You!',
-					description: 'Account created successfully',
+					description: 'Account updated successfully',
 					variant: 'default',
 				})
 				setTimeout(() => {
-					router.push('/')
+					router.replace('/account')
 				}, 1000)
 			}
 
@@ -115,7 +80,7 @@ const CreateAccountForm = () => {
 			if (error instanceof Error) {
 				toast({
 					title: 'Error',
-					description: 'An error occurred while creating the account',
+					description: 'An error occurred while updating the account',
 					variant: 'destructive',
 				})
 				console.error(error.message)
@@ -133,7 +98,7 @@ const CreateAccountForm = () => {
 	return (
 		<div className="mt-10 lg:mt-0">
 			<h2 className="text-lg md:text-xl lg:text-3xl mb-4 lg:mb-10">
-				Create Account
+				Edit Account
 			</h2>
 			<Form {...form}>
 				<form
@@ -222,7 +187,7 @@ const CreateAccountForm = () => {
 							</FormItem>
 						)}
 					/>
-					<FormField
+					{/* <FormField
 						control={form.control}
 						name="password"
 						render={({ field }) => (
@@ -241,14 +206,14 @@ const CreateAccountForm = () => {
 								</FormControl>
 								{isView ? (
 									<LuEye
-										className="absolute right-2 top-[34px] lg:top-[38px] z-10 cursor-pointer text-gray-500"
+										className="absolute right-2 top-8 z-10 cursor-pointer text-gray-500"
 										onClick={() => {
 											setIsView(!isView)
 										}}
 									/>
 								) : (
 									<LuEyeClosed
-										className="absolute right-2 top-[34px] lg:top-[38px] z-10 cursor-pointer text-gray-500"
+										className="absolute right-2 top-8 z-10 cursor-pointer text-gray-500"
 										onClick={() => setIsView(!isView)}
 									/>
 								)}
@@ -277,14 +242,14 @@ const CreateAccountForm = () => {
 								</FormControl>
 								{isViewConfirm ? (
 									<LuEye
-										className="absolute right-2 top-[34px] lg:top-[38px] z-10 cursor-pointer text-gray-500"
+										className="absolute right-2 top-8 z-10 cursor-pointer text-gray-500"
 										onClick={() => {
 											setIsViewConfirm(!isViewConfirm)
 										}}
 									/>
 								) : (
 									<LuEyeClosed
-										className="absolute right-2 top-[34px] lg:top-[38px] z-10 cursor-pointer text-gray-500"
+										className="absolute right-2 top-8 z-10 cursor-pointer text-gray-500"
 										onClick={() =>
 											setIsViewConfirm(!isViewConfirm)
 										}
@@ -293,15 +258,15 @@ const CreateAccountForm = () => {
 								<FormMessage className="text-red-500 text-sm md:text-base" />
 							</FormItem>
 						)}
-					/>
+					/> */}
 
 					<Button
 						disabled={isSubmitting}
 						type="submit"
 						variant="default"
-						className="text-sm md:text-base lg:text-lg bg-primary-light text-white hover:bg-primary-light shadow-md hover:shadow-xl hover:scale-[1.02] mt-8 px-6 py-2"
+						className="text-sm md:text-base lg:text-lg bg-primary-light text-white hover:bg-primary-light shadow-md hover:shadow-xl hover:scale-[1.02] mt-8 px-12 py-2"
 					>
-						Create Account
+						Save
 					</Button>
 				</form>
 			</Form>
@@ -309,4 +274,4 @@ const CreateAccountForm = () => {
 	)
 }
 
-export default CreateAccountForm
+export default EditAccountForm
