+++ src/components/account/send-offer-component.tsx
@@ -1,10 +1,10 @@
-import CreateAccountSummary from './create-account-summary'
-import SendOfferForm from './send-offer-form'
+import AccountSummaryFromStore from './summaries/account-summary-store'
+import SendOfferForm from './forms/send-offer-form'
 
 const SendOfferComponent = () => {
 	return (
 		<div className="w-4/5 mx-auto max-w-[500px] lg:grid grid-cols-2 lg:max-w-[1000px] lg:gap-20 text-primary mb-16 lg:mb-0">
-			<CreateAccountSummary />
+			<AccountSummaryFromStore />
 			<SendOfferForm />
 		</div>
 	)
